(Possibly) avoid compilation failure on mac

+++ Archipelago.cpp
@@ -69,7 +69,7 @@ void (*bouncedfunc)(AP_Bounce) = nullptr;
 // Serverdata Management
 std::map<std::string,AP_DataType> map_serverdata_typemanage;
 AP_GetServerDataRequest resync_serverdata_request;
-size_t last_item_idx = 0;
+uint64_t last_item_idx = 0;
 
 // Singleplayer Seed Info
 std::string sp_save_path;
